+++ src/module/sheets/AbstractTwodsixActorSheet.ts
@@ -10,6 +10,7 @@ import {onPasteStripFormatting} from "../sheets/AbstractTwodsixItemSheet";
 import { getKeyByValue } from "../utils/sheetUtils";
 import { resolveUnknownAutoMode } from "../utils/rollItemMacro";
 import { TWODSIX } from "../config";
+//import { applyEncumberedEffect } from "../hooks/showStatusIcons";
 
 export abstract class AbstractTwodsixActorSheet extends ActorSheet {
 
@@ -357,26 +358,6 @@ export abstract class AbstractTwodsixActorSheet extends ActorSheet {
     return this._onDropItemCreate(itemData);
   }
 
-  protected static _getWeight(item):number {
-    if ((item.type === "weapon") || (item.type === "armor") ||
-      (item.type === "equipment") || (item.type === "tool") ||
-      (item.type === "junk") || (item.type === "consumable")) {
-      if (item.system.equipped !== "ship") {
-        let q = item.system.quantity || 0;
-        const w = item.system.weight || 0;
-        if (item.type === "armor" && item.system.equipped === "equipped") {
-          if (item.system.isPowered) {
-            q = Math.max(0, q - 1);
-          } else {
-            q = Math.max(0, q - 1 + Number(game.settings.get("twodsix", "weightModifierForWornArmor")));
-          }
-        }
-        return (q * w);
-      }
-    }
-    return 0;
-  }
-
   protected static _prepareItemContainers(items, sheetData:TwodsixShipSheetData|any):void {
 
     // Initialize containers.
@@ -392,7 +373,6 @@ export abstract class AbstractTwodsixActorSheet extends ActorSheet {
     const spells:Item[] = [];
     const consumable:Item[] = [];
     const component = {};
-    let encumbrance = 0;
     let primaryArmor = 0;
     let secondaryArmor = 0;
     let radiationProtection = 0;
@@ -408,7 +388,6 @@ export abstract class AbstractTwodsixActorSheet extends ActorSheet {
         item.prepareConsumable();
       }
       if (sheetData.actor.type === "traveller") {
-        encumbrance += AbstractTwodsixActorSheet._getWeight(item);
         const anArmor = <Armor>item.system;
         if (item.type === "armor" && anArmor.equipped === "equipped") {
           primaryArmor += anArmor.armor;
@@ -476,18 +455,8 @@ export abstract class AbstractTwodsixActorSheet extends ActorSheet {
           break;
       }
     });
-    // Calc Max Encumbrance
-    let maxEncumbrance = 0;
-    if (sheetData.actor.type === "traveller") {
-      const encumbFormula = game.settings.get('twodsix', 'maxEncumbrance');
-      if (Roll.validate(encumbFormula)) {
-        maxEncumbrance = new Roll(encumbFormula, sheetData.actor.system).evaluate({async: false}).total;
-      } else {
-        ui.notifications.warn(game.i18n.localize("TWODSIX.Warnings.EncumbranceFormulaInvalid"));
-      }
-    }
 
-    // Assign and return sheetData.data to sheetData.system????
+    // Assign and return sheetData
     if (sheetData.actor.type === "traveller") {
       sheetData.container.equipment = equipment;
       sheetData.container.weapon = weapon;
@@ -502,8 +471,6 @@ export abstract class AbstractTwodsixActorSheet extends ActorSheet {
       sheetData.system.primaryArmor.value = primaryArmor;
       sheetData.system.secondaryArmor.value = secondaryArmor;
       sheetData.system.radiationProtection.value = radiationProtection;
-      sheetData.system.encumbrance.value = Math.round(encumbrance * 10) / 10; /*Round value to nearest tenth*/
-      sheetData.system.encumbrance.max = Math.round((maxEncumbrance || 0)* 10) / 10;
       sheetData.numberOfSkills = numberOfSkills + (sheetData.jackOfAllTrades > 0 ? 1 : 0);
       sheetData.skillRanks = skillRanks + sheetData.jackOfAllTrades;
     } else if (sheetData.actor.type === "animal" ) {